Rescrape Robert Christgau's albums
I realized that I hadn't scraped his Dean's lists, which are current up
to 2021. Most of those lists actually link to his reviews of the album.
What's more, there is a very flexible artist searching endpoint that
I can lazily link to that works shockingly well. I went ahead, wrote the
new scraper for the deans lists, fixed all the encoding issues I had
with that source, and updated the slugs to be links to links to his
actual writings!
