+++ app/config.ts
@@ -1,5 +1,6 @@
 const config = {
   siteTitle: 'Album Mode.party 🎉',
+  siteDescription: "Don't know what to listen to? Let us recommend an album!",
   cacheControl: {
     public: `public, max-age=${60 * 60 * 24}, s-maxage=${60 * 60 * 24}`,
     private: `private, max-age=${60 * 60 * 12}`,
