+++ app/lib/logging.server.ts
@@ -75,7 +75,9 @@ const logger = winston.createLogger({
   transports: [new winston.transports.Console()],
   format: winston.format.combine(
     winston.format.timestamp(),
-    winston.format.json()
+    winston.format.json({
+      space: env.SEED_SCRIPT ? 4 : undefined,
+    })
   ),
 })
